Treat channels as an LRU list
Freenode has a limit of 120 channels joined by a single connection.
Since some of the channels Gerritbot reports to are less active
than others, have it only join channels when it needs to send a
message.  If it has joined the max number of channels, part from
the least recently used channel to make room to join a new one.

Change-Id: Ifafeed77e640bc7b5910251c619358a8a2bd78a6