Don't autojoin channels
Don't pass the channels list to super in the irc bot class init method
because this appears to cause autojoining to happen which breaks because
we have too many channels. Instead let the event driven processing join
and part channels as needed to make the least recently used channel
rotation with channel limit work.

Change-Id: I447e75e84526d08fb9ec88eaae9a90f950dc5349